data/requirements.txt
polars
lxml_html_clean
typing_extensions
Scrapy
fastapi
uvicorn
pydantic
scikit-learn
requests
Flask
transformer
pandas
numpy
torch